encoding/json: Add SnytaxError to docu
SnytaxError was never mentioned in the doku. Browsing through the code,
I think that checkValid returns a SnytaxError if the scanner detects one.
As Unmarshal returns checkValid if the returned error is not null,
Unmarshal returns a SnytaxError, too, if checkValid detects one.

Therefore, I added a few sentences to the docu of SnytaxError, Unmarshal
and checkValid.

Change-Id: I71c9d9ef9d21a7ae9466d8c7b283fdfbba01f5a8