runtime/pprof: fix pcDeck's frame indexing
When building the inlining deck, correctly identify which is the last
frame in the deck. Otherwise, when some forms of inlining cause a PC to
expand to multiple frames, the length of the deck's two slices will
diverge.

Fixes the TestCPUProfileRecursion failures reported in CL 384239 (which
do not have an associated tracking issue).

Change-Id: I24e7ba32cb16b167f4307178b3f03c29e5362c4b