cmd/compile,go/build: remove embed import restriction for std
Some standard library packages can't easily import the embed package.
One such example is time/tzdata, as that results in an import cycle:

	time/tzdata -> embed -> io/fs -> context -> time -> time/tzdata

The import cycle is real, as the embed package uses fs.FS in its API.
However, it's very unfortunate in the case of time/tzdata,
because we really just want to embed a file as a string var.
We do not actually need to import the io/fs package at all.

A significant portion of standard library packages are in this position,
given that io/fs imports 27 packages in total.
There's really no good reason that those packages shouldn't be able to
use go:embed with string or []byte variables,
as that really only requires the toolchain to do the right thing.

An alternative approach that was considered was "internal/embedlite",
an empty sentinel package that could be imported instead of "embed",
meaning that std packages could satisfy go:embed's import restriction
without having to pull in io/fs as a dependency.

Either approach accomplishes the same result without changing the import
restriction for packages outside of the standard library.
We chose the one that doesn't introduce an extra package,
as that feels easier to implement and maintain in the long run.
Plus, we don't need to teach go/build/deps_test.go about the new package.

Fixes #51463.

Change-Id: I233081276d5c4d718d183a84db68298130be0cc3