runtime: add logging in TestCtrlHandler and skip it
I tried adding logging for this test to figure out why it is failing
when run directly on a gomote. The logging helps, but it doesn't fix
the failure, and the persistent failure is impeding my ability to test
other changes for #51461.

Let's skip it until it can be diagnosed.

Updates #51602.
Updates #41884.

Change-Id: I044f42c36697afbbba1174edb1fc53dd0d898a21