runtime: skip TestCtrlHandler if taskkill fails to terminate it gracefully
It might not be possible to send a CTRL_CLOSE_EVENT to a process
running under a pseudoconsole,¹ as is the case when connected to
Windows via OpenSSH. Ideally we should skip the test entirely
if running under a pseudoconsole, but I couldn't find a way to
detect that condition upfront. Instead, we'll build the binary
and start the subprocess, but skip the test (and clean up the
subprocess) if taskkill fails to terminate it gracefully.

I also noticed in the course of debugging that this test leaks
the child process if the test times out. To attempt to avoid
that leak, I have added a deadline timer that forcefully cancels
it before the test times out.

¹https://docs.microsoft.com/en-us/windows/console/creating-a-pseudoconsole-session

Updates #51602.
Updates #41884.

Change-Id: I044f42c36697afbbba1174edb1fc53dd0d898a21