runtime: disable idle mark workers during periodic GC cycles
This change disables idle mark workers during periodic (currently every
2 minutes) GC cycles that trigger in applications that are idle with
respect to the GC (i.e. don't allocate much).

Idle mark workers soak up all available and unused Ps, up to GOMAXPROCS.
While this provides some throughput and latency benefit in general, it
can cause what appear to be massive CPU utilization spikes in otherwise
idle applications. This is mostly an issue for *very* idle applications,
ones idle enough to trigger periodic GC cycles. This spike also tends to
interact poorly with auto-scaling systems, as the system might assume
the load average is very low and suddenly see a massive burst in
activity.

The result of this change is not to bring down this 100% (of GOMAXPROCS)
CPU utilization spike to 0%, but rather to 25%, the amount of CPU time
the GC dedicates to background mark workers.

Idle mark workers also do incur a small latency penalty as they must be
descheduled for other work that might pop up. Luckily the runtime is
pretty good about getting idle mark workers off of Ps, so in general
the latency benefit from shorter GC cycles outweighs this cost. But, the
cost is still non-zero and may be more significant in idle applications
that aren't invoking assists and write barriers quite as often.

Fixes #44163.

Change-Id: Ib74793bb6b88d1765c52d445831310b0d11ef423