runtime: check the heap goal and trigger dynamically
As it stands, the heap goal and the trigger are set once by
gcController.commit, and then read out of gcController. However with the
coming memory limit we need the GC to be able to respond to changes in
non-heap memory. The simplest way of achieving this is to compute the
heap goal and its associated trigger dynamically.

In order to make this easier to implement, the GC trigger is now based
on the heap goal, as opposed to the status quo of computing them
simultaneously. In many cases we just want the heap goal anyway, not
both, but we definitely need the goal to compute the trigger, because
the trigger's bounds are entirely based on the goal (the initial runway
is not). A consequence of this is that we can't rely on the trigger to
enforce a minimum heap size anymore, and we need to lift that up
directly to the goal. Specifically, we need to lift up any part of the
calculation that *could* put the trigger ahead of the goal. Luckily this
is just the heap minimum and minimum sweep distance. In the first case,
the pacer may behave slightly differently, asthe heap minimum is no longer
the minimum trigger, but the actual minimum heap goal. In the second
case it should be the same, as we ensure the additional runway for
sweeping is added to both the goal *and* the trigger.

Another consequence of this is that there's a place where the heap goal
is altered if we triggered too late (for one reason or another, usually
due to external factors we can't control). Currently it uses the
computed trigger, but because the trigger now depends on the heap goal,
we can't actually make that check as it exists. So, this change also
introduces a "triggered" field into gcControllerState which tracks the
point that we *actually* triggered. In fact, there are quite a few
places where we really intend to talk about when we actually triggered,
but instead use the computed trigger, so this change also uses
"triggered" in those places.

This all sounds like a lot that changed, but what we're talking about
here are really, really tight corner cases that arise from situations
outside of our control, like pathologically bad behavior on the part of
an OS or CPU. In practice, nothing much has really changed here.

In fact, one may at first think that computing the heap goal and trigger
dynamically introduces some raciness, but not in this CL: the heap goal
and trigger are still almost completely static (aside from the min sweep
distance corner case).

Allocation outside of a GC cycle may now be a bit slower than before, as
the GC trigger check is now significantly more complex. However, note
that this executes basically just as often as gcController.revise, and
that makes up for a vanishingly small part of any CPU profile. The next
CL cleans up the floating point multiplications on this path
nonetheless, just to be safe.

For #48409.

Change-Id: I280f5ad607a86756d33fb8449ad08555cbee93f9