runtime: check the heap goal and trigger dynamically
As it stands, the heap goal and the trigger are set once by
gcController.commit, and then read out of gcController. However with the
coming memory limit we need the GC to be able to respond to changes in
non-heap memory. The simplest way of achieving this is to compute the
heap goal and its associated trigger dynamically.

There are many implications to this behavior, but for the sake of this
change, they shouldn't matter. Ultimately, the inputs to the heap goal
are static as before, so really this change is a no-op refactoring.
Allocation outside of a GC cycle may be a little bit slower than before
as the GC trigger check is now much more complicated. The next CL cleans
up the floating point multiplications on that path.

For #48409.

Change-Id: I280f5ad607a86756d33fb8449ad08555cbee93f9